Warn users that config file must be named 'reviewers' for now
Due to a bug in Gerrit core, that was fixed in I9507de65a config file
that named as documented: 'reviewers.config' is not found. Add a warning
in the documentation that untill the fixed Gerrit version is released
(2.10) the config file for this plugin must be renamed to 'reviewers'
to be found by Gerrit core.

Reported-by: Alexander Smirnov <alexander@smirn0v.ru>
Change-Id: Ifc0751a12f0ec8f9e5690deb7c6268a34ae722b4