Remove VERSION
Hard-coding the version makes it difficult to know what version
of the plugin is running, because the version displayed will be
the same regardless of which revision it was built from when
building with buck.

Remove the VERSION, so that the version of the plugin is taken
from the git describe, and thus we can see exactly what version
is running.

Change-Id: I4dff6eef01a7be5599caf8443690c05ba365b30b