Fix wrong date/time for commits in refs/meta/config branch
When the refs/meta/config branch is modified using the PutReviewers
REST endpoint, the commit date/time is wrong. Instead of the actual
date/time the date/time of the last Gerrit server start is used.

This is because MetaDataUpdate.User which gets the GerritPersonIdent
injected is kept as member in the PutReviewers singleton and the
date/time for commits in the refs/meta/config branch is retrieved from
that GerritPersonIdent instance which is only created once when the
PutReviewers singleton is instantiated.

Change-Id: I851522efa1e4117c5ffcb2741bf33079d4fb5495